列車コンテナのセーブ、ロードの仕組みのJsonObjectの基盤実装
